% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{abs_madm}
\alias{abs_madm}
\title{MADM}
\usage{
abs_madm(pred)
}
\arguments{
\item{pred}{T X N Matrix of predictions. Each column is
a simulation.}
}
\value{
vector of length T.
}
\description{
Median absolute deviation about the median
}
\details{
\deqn{median(|pred - median(pred)|)}

Median absolute deviation about the median is a measure of
how clustered the forecasts are. A value of 0 indicates that all
the predicted values are the same, thus highly clustered. Large
values indicate more diffuse predictions.
}
\references{
https://bit.ly/2vPO0I9
}
\seealso{
\code{\link[=rel_madm]{rel_madm()}}
}
