% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{abs_median_dvtn}
\alias{abs_median_dvtn}
\title{Sharpness}
\usage{
abs_median_dvtn(pred)
}
\arguments{
\item{pred}{T X N Matrix of predictions. Each column is
a simulation.}
}
\value{
vector of length T.
}
\description{
Absolute Median absolute deviation about the median
}
\details{
\deqn{median(|pred - median(pred)|)}
}
\references{
https://bit.ly/2vPO0I9
}
