% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{all_col_vals_unique}
\alias{all_col_vals_unique}
\title{Are all column values unique?}
\usage{
all_col_vals_unique(subtbl, subcol, add = TRUE)
}
\arguments{
\item{subtbl}{name of submission table}

\item{subcol}{name of submission column}
}
\value{
logical
}
\description{
Are all column values unique?
}

