% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apa_p}
\alias{apa_p}
\title{Report a p-value in APA format}
\usage{
apa_p(x)
}
\arguments{
\item{x}{The p-value.}
}
\value{
A string with formatted p-value.
}
\description{
Report a p-value in APA format
}
