% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apa_t}
\alias{apa_t}
\title{Report a t-test in APA format}
\usage{
apa_t(x)
}
\arguments{
\item{x}{A t-test object, result of the call to \code{t.test}.}
}
\value{
A string displaying results in APA format.
}
\description{
Report a t-test in APA format
}
