% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{apply_blacklist}
\alias{apply_blacklist}
\title{Remove blacklisted functions}
\usage{
apply_blacklist(subfile)
}
\arguments{
\item{subfile}{path to submission RMarkdown file}
}
\value{
a tibble with columns \code{blacklist}, whether any functions were found; and \code{bfns}, names of these functions
}
\description{
Comment out any blacklisted functions that appear within a file
}
\details{
warning: the original file will be overwritten, with any
  lines containing blacklisted code commented out
}
\seealso{
\code{\link{get_blacklist}}
}

