% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_columns_identical}
\alias{are_columns_identical}
\title{Are table column values identical}
\usage{
are_columns_identical(subtbl, subcol, sol_env, soltbl = subtbl,
  solcol = subcol, ignore_order = FALSE, add = TRUE)
}
\arguments{
\item{subtbl}{name of submission table}

\item{subcol}{name of column in submission table}

\item{sol_env}{solution environment}

\item{soltbl}{name of table of solution environment}

\item{solcol}{name of column in solution table}

\item{ignore_order}{should the ordering of the values be ignored?}

\item{add}{whether to add feedback}
}
\value{
logical
}
\description{
Are table column values identical
}
