% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{are_num_vals_close}
\alias{are_num_vals_close}
\title{Are submission and solution numeric values close}
\usage{
are_num_vals_close(valname, sol_env, ignore.case = FALSE, tolerance = 0.002,
  add = TRUE)
}
\arguments{
\item{valname}{Name of the variable whose value you want to test}

\item{sol_env}{solution environment}

\item{ignore.case}{whether to accept same variable name but different capitalization}

\item{tolerance}{how close the values have to be}

\item{add}{whether to add feedback}
}
\value{
logical; \code{TRUE} if \code{abs(x - get(x, sol_env)) < tolerance}
}
\description{
Test whether numeric values in the submission are close to corresponding values in the solution environment.
}

