% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{are_tables_identical}
\alias{are_tables_identical}
\title{Are submission and solution tables identical}
\usage{
are_tables_identical(tblname, sol_env, ignore.case = FALSE)
}
\arguments{
\item{tblname}{name of the table}

\item{sol_env}{the solution environment}

\item{ignore.case}{whether to ignore case when matching submission and solution variable names}
}
\value{
logical; returns result of \code{dplyr::setequal(tblname, get(tblname, sol_env))}
}
\description{
Are submission and solution tables identical
}

