% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess_all}
\alias{assess_all}
\title{Assess all files in a directory}
\usage{
assess_all(
  dirname,
  key,
  sub_id = moodle_id(list_submissions(dirname)),
  use_sub_env = TRUE,
  workdir = NULL,
  seed = NULL,
  preseed = NULL,
  stop_after = -1L,
  task_varlist = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dirname}{name of the directory containing the submission files}

\item{key}{Must be either: (1) A single key with answers to the problems, normally result of call to \code{link{compile_key}}; or (2) A named list of keys, with names matching the values of \code{sub_id}.}

\item{sub_id}{subject identifies (vector same length as \code{dirname})}

\item{use_sub_env}{process submission code in the submission environment (\code{FALSE} to process it in the solution environment)}

\item{workdir}{working directory (deprecated as of version 0.2.0.0000)}

\item{seed}{random seed to set at the beginning}

\item{preseed}{Named list with random seed to set before each (named) block.}

\item{stop_after}{stop processing after completing N files}

\item{task_varlist}{Named list of lists, with each element having names corresponding to the names in \code{sub_id}, and with each sublist being itself a named list whose names correspond to task chunks. Elements of that second list should be named character vectors with task variables.}

\item{verbose}{Whether to report on progress assessing individual tasks within assessments.}
}
\value{
a dataframe with the assessment variables and their values
}
\description{
Convenience function that runs the function \code{assess} on all files within the directory specified by \code{dirname}.
}
