% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assess.R
\name{assess_submission}
\alias{assess_submission}
\title{Assess a submission}
\usage{
assess_submission(filename, sub_id = filename, key)
}
\arguments{
\item{filename}{name of submission file}

\item{sub_id}{unique submission identifier}

\item{key}{key table (see \code{\link{compile_key}})}
}
\value{
A table
}
\description{
Assess a submission
}

