% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assess.R
\name{assessment_code}
\alias{assessment_code}
\title{Create an assessment code file}
\usage{
assessment_code(s_file, o_file = "assess_code.Rmd", overwrite = FALSE)
}
\arguments{
\item{s_file}{name of file with solutions}
}
\value{
name of the output file
}
\description{
Create an assessment code file based on solution
}

