% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{attempted}
\alias{attempted}
\title{Check whether any attempt was made to answer the question}
\usage{
attempted(subvar, add = TRUE, inherits = FALSE)
}
\arguments{
\item{subvar}{Name of the submission variable to test.}
}
\value{
Returns \code{FALSE} only if \code{subvar} is \code{NULL}.
}
\description{
Check whether any attempt was made to answer the question
}
\details{
If no attempt was made, then the value of \code{subvar} will remain \code{NULL}.
}
