% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{avg_residual}
\alias{avg_residual}
\title{Average Residual}
\usage{
avg_residual(obs, pred)
}
\arguments{
\item{obs}{observed vector T X 1}

\item{pred}{matrix of predicted observations. Each column is a
simulation. T X N where N is the number of simulations.}
}
\value{
error T X 1. Each entry is the error averaged across the
simulations.
}
\description{
Resdiual averaged acorss simulations
}
\details{
\deqn{\sum_{i = 1}^{N}{obs - pred} / N }
}
\author{
Sangeeta Bhatia
}
