% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{bias}
\alias{bias}
\title{Bias}
\usage{
bias(obs, pred)
}
\arguments{
\item{obs}{observed vector T X 1}

\item{pred}{Simulated predictions T X N. Each column is a simulation.}
}
\value{
vector of length T.
}
\description{
Bias in probabilistic forecasts
}
\details{
Bias is measured as
\deqn{2 * mean((heaviside(obs - pred)) - 0.5)}
where heaviside returns 1 if the arg is positive, 0 if this negative
and 0.5 if it is 0. The average is taken over all simulations.
}
\references{
https://bit.ly/2vPO0I9
}
\author{
Sangeeta Bhatia
}
