% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{boot_ci}
\alias{boot_ci}
\title{Calculate bootstrapped confidence intervals}
\usage{
boot_ci(FUN = NULL, nboot = 100, alpha = 0.05, na.rm = FALSE, ...)
}
\arguments{
\item{FUN}{Function to bootstrap. Must return a single value.}

\item{nboot}{Default 100. Number of iterations to use to estimate
the output statistic confidence interval.}

\item{alpha}{Default 0.05. Numeric value indicating the confidence
interval to return. 0.05 will return the 95\% confidence interval.}

\item{na.rm}{Default FALSE. A boolean value indicating whether or not to
remove NA values. If missing values are present but not removed the
function will output NA.}

\item{...}{Named arguments passed on to \code{FUN}.}
}
\value{
A two-long numeric vector containing the bootstrapped confidence
  interval of the input vector(s).
}
\description{
Calculate the non-parametric bootstrap confidence interval
  for a given numeric input and a chosen function.
}
\examples{

# Calculate COD confidence interval
boot_ci(cod, nboot = 100, ratio = ratios_sample$ratio)

# Calculate PRD confidence interval
boot_ci(
  prd,
  nboot = 100,
  assessed = ratios_sample$assessed,
  sale_price = ratios_sample$sale_price,
  na.rm = FALSE
)
}
