% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_assessment}
\alias{browse_assessment}
\title{Browse assessment}
\usage{
browse_assessment(x)
}
\arguments{
\item{x}{filename of the assessment results \code{.rds} file or the tibble contained in such file}
}
\value{
table containing the new values and feedback for the task
}
\description{
View assessment code and figure output and modify assessment values
in a shiny gadget
}
\details{
launches a shiny gadget allowing the assessor to change the auto-generated values
}

