% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_inline}
\alias{browse_inline}
\title{Browse inline responses in shiny app}
\usage{
browse_inline(
  htbl,
  start_delim = "START HERE.*",
  end_delim = "END HERE.*",
  outfile = "inline_scores.rds",
  overwrite = FALSE
)
}
\arguments{
\item{htbl}{Data frame with two columns, \code{sub_id} (submission id) and \code{filename}, name of html file.}

\item{start_delim}{Delimiter for html comment starting text block (regular expression)}

\item{end_delim}{Delimiter for html comment ending text block (regular expression)}

\item{outfile}{Name of output file to write response data to.}

\item{overwrite}{Whether to overwrite the output file if it already exists.}
}
\description{
Browse inline responses in shiny app
}
