% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdu_dict}
\alias{cdu_dict}
\title{Codes used by the CCAO to identify certain distinct property situations.}
\format{
A data frame with 28 rows and 4 variables:
\describe{
  \item{cdu_code}{Common name of the township}
  \item{cdu_type}{Two-digit code used to identify the township}
  \item{cdu_value}{Actual meaning of CDU code}
  \item{cdu_value_short}{Actual meaning of CDU code, shortened}
}
}
\usage{
cdu_dict
}
\description{
A dataset containing a lookup of CDU codes. These codes are kind of a mess
and have been created and used inconsistently over the years.
}
\keyword{datasets}
