% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chars_dict}
\alias{chars_dict}
\title{Data dictionary for CCAO characteristic values.}
\format{
A data frame with 61 rows and 4 variables:
\describe{
  \item{char_name}{Column name of variable in CCAO SQL server}
  \item{char_code}{Value code used in CCAO SQL server and in paper forms}
  \item{char_value}{Actual meaning of char_code}
  \item{char_value_short}{Actual meaning of char_code, shortened}
}
}
\source{
This dictionary was manually created from paper forms as a
  translation of numeric variables. char_value_short is the equivalent of
  what is used on the AS400 property info screens
}
\usage{
chars_dict
}
\description{
A dataset containing human-readable version of numeric characteristic
encodings of property characteristic data.
}
\keyword{datasets}
