% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{chk}
\alias{chk}
\title{Checkmark, dash, or x depending on logical value}
\usage{
chk(x)
}
\arguments{
\item{x}{Logical value}
}
\value{
HTML with the code for a green checkmark or red X depending on the value of 'x'.
}
\description{
Checkmark, dash, or x depending on logical value
}
