% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junk.R, R/print.R
\name{chk_num}
\alias{chk_num}
\title{Build a table showing numeric values against solution}
\usage{
chk_num(vnames, results, nums, solenv, format = "markdown", digits = 3)

chk_num(vnames, results, nums, solenv, format = "markdown", digits = 3)
}
\arguments{
\item{vnames}{Character vector of names of variables to be checked.}

\item{results}{List of results from \code{\link{num_vals_close}}}

\item{nums}{Vector of numbers from solution (NAs if missing).}

\item{solenv}{Solution environment.}

\item{format}{Format of the output (argument passed to  \code{knitr::kable}).}

\item{digits}{Number of digits for formatting output.}

\item{x}{Character vector of names of variables to be checked.}

\item{tolerance}{Tolerance for call to \code{\link{num_vals_close}}.}
}
\description{
Build a table showing numeric values against solution

Build a table showing numeric values against solution
}
\details{
Checking performed with `code{\link{attempted}}` and `\code{\link{num_vals_close}}`.

Checking performed with `code{\link{attempted}}` and `\code{\link{num_vals_close}}`.
}
