% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class_dict}
\alias{class_dict}
\title{Data dictionary of Cook County property classes.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{class_code}{Class code, numeric}
  \item{desc}{Human-readable description of the property class}
}
}
\usage{
class_dict
}
\description{
A dataset containing a translation for residential class codes to
human-readable class descriptions.
}
\note{
Only includes residential classes.
}
\keyword{datasets}
