% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R, R/ci.R
\name{cod}
\alias{cod}
\alias{cod_ci}
\alias{cod_met}
\title{Calculate Coefficient of Dispersion (COD)}
\usage{
cod(estimate, sale_price, na.rm = FALSE)

cod_ci(estimate, sale_price, nboot = 1000, alpha = 0.05, na.rm = FALSE)

cod_met(x)
}
\arguments{
\item{estimate}{A numeric vector of assessed values. Must be the same
length as \code{sale_price}.}

\item{sale_price}{A numeric vector of sale prices. Must be the same length
as \code{estimate}.}

\item{na.rm}{Default FALSE. A boolean value indicating whether or not to
remove NA values. If missing values are present but not removed the
function will output NA.}

\item{nboot}{Default 1000. Number of iterations to use to estimate
the output statistic confidence interval.}

\item{alpha}{Default 0.05. Numeric value indicating the confidence
interval to return. 0.05 will return the 95\% confidence interval.}

\item{x}{Numeric vector of sales ratio statistic(s) to check
against IAAO/Quintos standards.}
}
\description{
COD is the average absolute percent deviation from the
  median ratio. It is a measure of horizontal equity in assessment.
  Horizontal equity means properties with a similar fair market value
  should be similarly assessed.

  Lower COD indicates higher uniformity/horizontal equity in assessment.
  The IAAO sets uniformity standards that define generally accepted ranges
  for COD depending on property class. See
  \href{https://www.iaao.org/media/standards/Standard_on_Ratio_Studies.pdf}{IAAO Standard on Ratio Studies}
  Section 9.1, Table 1.3 for a full list of standard COD ranges.

  NOTE: The IAAO recommends trimming outlier ratios before calculating COD,
  as it is extremely sensitive to large outliers. The typical method used is
  dropping values beyond 3 * IQR (inner-quartile range). See
  \href{https://www.iaao.org/media/standards/Standard_on_Ratio_Studies.pdf}{IAAO Standard on Ratio Studies}
  Appendix B.1.
}
\section{Functions}{
\itemize{
\item \code{cod()}: Returns a numeric vector containing the COD of \code{ratios}.

\item \code{cod_ci()}: Returns upper and lower CI as a named vector.

\item \code{cod_met()}: Returns TRUE when input COD meets IAAO standards
(between 5 and 15).

}}
\examples{
# Calculate COD
cod(ratios_sample$estimate, ratios_sample$sale_price)

# Calculate COD confidence interval
cod_ci(ratios_sample$estimate, ratios_sample$sale_price)
}
\seealso{
Other formulas: 
\code{\link{mki}()},
\code{\link{prb}()},
\code{\link{prd}()}
}
\concept{formulas}
