% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaao_func.R
\name{cod_func}
\alias{cod_func}
\title{Calculate the COD of a given set of assessment ratios}
\usage{
cod_func(
  ratios,
  trim = c(0.05, 0.95),
  bootstrap_n = 100,
  suppress = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{ratios}{A vector of ratios centered around 1, where the numerator of
the ratio is the estimated fair cash value and the denominator
is the sale price (sometimes lagged or leading).}

\item{trim}{A lower and upper bound for trimming assessment ratios. These
values are set by CCAO Data Science Department SOPs. Do not alter
them without permission.}

\item{bootstrap_n}{The number of iterations to use to estimate standard error
and 95\% confidence interval. Setting to 0 or FALSE will not use bootstapping.}

\item{suppress}{Default FALSE. If TRUE, hide warning about not meeting the
minimum number of ratios to calculate COD accurately dplyr (30 by default).}

\item{na.rm}{Default FALSE. If TRUE, will remove NAs from the vector before
proceeding.}
}
\value{
A named list containing COD, standard error, 95\% confidence interval,
  and the number of ratios used in the calculation
}
\description{
Calculate the COD of a given set of assessment ratios
}
\examples{
library(assessR)

ratios <- runif(300, 0.93, 1.08)

cod_func(ratios, trim = c(0.05, 0.95), bootstrap_n = 100)

}
