% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{compile_all}
\alias{compile_all}
\title{Compile all RMarkdown submissions}
\usage{
compile_all(dir, quiet = TRUE, blacklist = TRUE, progress = TRUE,
  with_moodle_id = FALSE, knit_root_dir = NULL)
}
\arguments{
\item{dir}{directory containing RMarkdown submissions}

\item{quiet}{suppress messages while compiling}

\item{blacklist}{comment out lines of code containing blacklisted functions}

\item{progress}{report file-by-file progress}

\item{knit_root_dir}{root directory for knitting (see \code{\link{rmarkdown::render}})}

\item{return}{should return table contain moodle id number}
}
\value{
table with information about compilation for each submission
}
\description{
Compile all RMarkdown submissions
}
\details{
Convenience function to call \code{\link{compile_assignment}} repeatedly for each RMarkdown file in subdirectory \code{subdir}
}
