% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{compile_key}
\alias{compile_key}
\title{Compile the assessment key}
\usage{
compile_key(s_file, a_file, overwrite = FALSE, workdir = NULL)
}
\arguments{
\item{s_file}{name of Rmd solution file}

\item{a_file}{name of Rmd file with assessment code}

\item{overwrite}{overwrite \code{solution} directory if it exists?}

\item{workdir}{working directory}
}
\value{
a list containing the starting environments for each task
}
\description{
Compile the assessment key, creating starting and solution
environments to use in assessment, as well as output to use in
feedback reports.
}
\details{
Sets up environments for running submission code and
  assessment code.  Creates object \code{sol_env} in each
  environment which has the solution environment needed for
  assessment.  Also creates output/plots when appropriate and
  stores them in the \code{solution} subdirectory in the current
  working directory.
}
\seealso{
\code{\link{tangle}}
}
