% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{copy_data}
\alias{copy_data}
\title{Copy data to RMarkdown submission directory}
\usage{
copy_data(subdir, datadir)
}
\arguments{
\item{subdir}{subdirectory of submission files}

\item{datadir}{subdirectory of data files}
}
\description{
Copy data to RMarkdown submission directory
}
\details{
copies over data from \code{datadir} to each submission subdirectory
}

