% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{def_by_code}
\alias{def_by_code}
\title{Test whether a variable was defined statically or using code}
\usage{
def_by_code(x, code, ignore_case = FALSE, add = TRUE)
}
\arguments{
\item{x}{name of the variable to check}

\item{ignore_case}{whether to ignore case of \code{x}}

\item{add}{whether to add feedback}
}
\value{
logical; \code{TRUE} if the code defining the variable uses any function
}
\description{
Test whether a variable was defined statically or using code
}

