% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_comment.R
\name{extract_html_comment}
\alias{extract_html_comment}
\title{Extract html comment}
\usage{
extract_html_comment(filename, dbegin_regex, dend_regex = dbegin_regex,
  remove_leading_blanks = TRUE, remove_trailing_blanks = TRUE)
}
\arguments{
\item{filename}{name of the html or Rmd file}

\item{dbegin_regex}{delimiter at start of block (regular expression)}

\item{dend_regex}{delimiter at end of block (regular expression)}

\item{remove_leading_blanks}{whether to keep blank lines at the start of the content}

\item{remove_ending_blanks}{whether to keep blank lines at the end of the content}
}
\value{
character vector with the lines between the delimiters
}
\description{
Extract an html comment block from a file.
}
\details{
Will fail if either: there is more than one comment block matching either delimiter; or the ending delimiter occurs before the start delimiter.
}
