% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{feedback_all}
\alias{feedback_all}
\title{Produce feedback reports}
\usage{
feedback_all(ares, template, subdir = "feedback_reports", overwrite = FALSE,
  quiet = TRUE, extra_params = NULL, long_line_cutoff = 75,
  empty_fbk = "* no issues", stop_after = -1L)
}
\arguments{
\item{ares}{assessment result}

\item{template}{an RMarkdown template for the report}

\item{subdir}{the subdirectory for reports}

\item{overwrite}{if exists, overwrite?}

\item{quiet}{'quiet' option for rmarkdown::render}

\item{extra_params}{any extra parameters to pass to report (will appear in params$extra)}

\item{long_line_cutoff}{reformat chunks where any single line is longer than this value (-1 = don't reformat)}

\item{empty_fbk}{default feedback (when \code{fbk} field is empty)}

\item{stop_after}{stop processing after N (-1 to process all)}
}
\value{
vector of report filenames
}
\description{
Produce feedback reports
}
