% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\name{feedback_report}
\alias{feedback_report}
\title{Produce a feedback report}
\usage{
feedback_report(d, template, subdir = "feedback_reports", overwrite = FALSE,
  quiet = TRUE)
}
\arguments{
\item{d}{a table with assessment data for a single submission}

\item{template}{an RMarkdown template for the report}

\item{subdir}{the subdirectory for reports}

\item{overwrite}{if exists, overwrite?}

\item{quiet}{'quiet' option for rmarkdown::render}
}
\value{
path to the file
}
\description{
Produce a feedback report
}

