% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{feedback_template}
\alias{feedback_template}
\title{Create a feedback template for the assessment}
\usage{
feedback_template(
  a_file,
  s_file,
  o_file = "feedback_template.Rmd",
  overwrite = FALSE
)
}
\arguments{
\item{a_file}{rmarkdown file with assessment code}

\item{s_file}{rmarkdown file with solutions}

\item{o_file}{name of output file}

\item{overwrite}{overwrite the file if it exists}
}
\value{
name of the output file
}
\description{
Create a feedback template for the assessment
}
