% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{fetch}
\alias{fetch}
\title{Download assignments from moodle}
\usage{
fetch(assign_id, targ_dir = paste0("assign_", assign_id),
  overwrite = FALSE, url = NULL, user = NULL)
}
\arguments{
\item{assign_id}{assignment id number; digits only}

\item{targ_dir}{subdirectory in which to store downloaded submissions; if \code{NULL} will be downloaded to a subdirectory of the working directory named \code{assign_id}}

\item{overwrite}{overwrite the subdirectory if it exists (default \code{FALSE})}

\item{url}{URL of moodle server (if not specified in credential file)}

\item{user}{name of moodle user (if not specified in credential file)}
}
\value{
a tibble containing information about contents of the downloaded directory (see \code{\link{list_submissions}})
}
\description{
\code{fetch} downloads student submissions for a given assignment from a moodle server.
}
\seealso{
\code{\link{login}}, \code{\link{rmdassess::describe}}
}
