% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_exists}
\alias{fun_exists}
\title{Does function exist in submission environment?}
\usage{
fun_exists(fnname, add = TRUE)
}
\arguments{
\item{fnname}{name of function}

\item{add}{add feedback? (logical)}
}
\value{
logical
}
\description{
Does function exist in submission environment?
}
