% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_result_identical}
\alias{fun_result_identical}
\title{Do return values from a function match solution?}
\usage{
fun_result_identical(subfn, args = list(), solenv, solfn = subfn,
  sameseed = TRUE, add = TRUE)
}
\arguments{
\item{subfn}{name of function in submission environment}

\item{args}{list of args to pass to function}

\item{solenv}{solution environment}

\item{solfn}{name of solution function}

\item{sameseed}{whether to run the two functions with the same state of \code{.Random.seed}}

\item{add}{whether to add feedback}
}
\value{
logical
}
\description{
Do return values from a function match solution?
}
\details{
compare the result of the function in the submission environment to its result in the solution environment
}

