% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fun_try}
\alias{fun_try}
\title{Safely try out a function defined in the submission}
\usage{
fun_try(fnname, ..., add = TRUE)
}
\arguments{
\item{fnname}{name of the function}

\item{...}{arguments to fnname}

\item{add}{whether to add feedback}
}
\value{
list with two values: \code{is_error} (logical) and \code{result}
}
\description{
Safely try out a function defined in the submission
}
