% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_err_string}
\alias{get_err_string}
\title{Get a printable string from error message}
\usage{
get_err_string(x)
}
\arguments{
\item{x}{a condition}
}
\value{
string with the error message
}
\description{
Get a printable string from error message
}
