% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%==\%}
\alias{\%==\%}
\title{Vectorized version of all.equal}
\usage{
x \%==\% y
}
\arguments{
\item{x}{vector whose elements you want to compare}

\item{y}{vector whose elements you want to compare to x}
}
\description{
Vectorized version of all.equal
}
\details{
performs element-by-element comparison, dealing properly
  with floating-point values
}
