% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_columns}
\alias{has_columns}
\title{Does a submission table have required columns?}
\usage{
has_columns(subtbl, subcols, add = TRUE)
}
\arguments{
\item{subtbl}{name of table in submission environment}

\item{subcols}{names of columns that should exist}

\item{add}{whether to give feedback}
}
\value{
logical
}
\description{
Does a submission table have required columns?
}
