% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{heaviside}
\alias{heaviside}
\title{Heaviside function}
\usage{
heaviside(x)
}
\arguments{
\item{x}{number}
}
\value{
1 if x is positive, 0 if it is negative and 0.5 otherwise.
}
\description{
Heaviside function
}
\details{
Heaviside function H(x) is defined to be 1 if x is positive
and 0 otherwise. If x is 0, the function returns 0.5.
}
\author{
Sangeeta Bhatia
}
\keyword{internal}
