% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales_chasing.R
\name{is_sales_chased}
\alias{is_sales_chased}
\title{Detect sales chasing in a vector of sales ratios}
\usage{
is_sales_chased(
  x,
  method = "both",
  bounds = c(0.98, 1.02),
  gap = 0.03,
  na.rm = FALSE
)
}
\arguments{
\item{x}{A numeric vector. Must be longer than 2 and cannot contain
\code{inf} or \code{NA} values.}

\item{method}{Default \code{both}. String indicating sales chasing detection
method. Options are \code{cdf}, \code{dist}, or \code{both}.}

\item{bounds}{Default \code{(0.98, 1.02)}. Lower and upper bounds of the
range of ratios to consider when detecting sales chasing. Setting this to
a narrow band at the center of the ratio distribution prevents detecting
false positives at the tails.}

\item{gap}{Default \code{0.05}. Float tuning factor. For the CDF method, it
sets the maximum percentage difference between two adjacent ratios. For the
distribution method, it sets the maximum percentage point difference
between the percentage of the data between the \code{bounds} in the real
distribution compared to the ideal distribution.}

\item{na.rm}{Default FALSE. A boolean value indicating whether or not to
remove NA values. If missing values are present but not removed the
function will output NA for those values.}
}
\value{
A logical value indicating whether or not the input values may
  have been sales chased.
}
\description{
Sales chasing is when a property is selectively reappraised to
  shift its assessed value toward its recent sale price. Sales chasing is
  difficult to detect. This function is NOT a statistical test and does
  not provide the probability of the given result. Rather, it combines two
  heuristic methods to roughly estimate if sales chasing has occurred.

  The first method (cdf) detects discontinuities in the cumulative
  distribution function (CDF) of the input vector. Ratios that are not sales
  chased should have a fairly smooth CDF. Discontinuous jumps in the CDF,
  particularly around 1, may indicate sales chasing. This can usually be seen
  visually as a "flat spot" on the CDF.

  The second method (dist) uses the technique outlined in the
  \href{https://www.iaao.org/media/standards/Standard_on_Ratio_Studies.pdf}{IAAO Standard on Ratio Studies}
  Appendix E, Section 4. It compares the percentage of real data within +-2%
  of the mean ratio to the percentage of data within the same bounds given a
  constructed normal distribution with the same mean and standard deviation.
  The intuition here is that ratios that are sales chased may be more
  "bunched up" in the center of the distribution.
}
\examples{

# Generate fake data with normal vs chased ratios
normal_ratios <- c(rnorm(1000, 1, 0.15))
chased_ratios <- c(rnorm(900, 1, 0.15), rep(1, 100))

# Plot to view discontinuity
plot(stats::ecdf(normal_ratios))
is_sales_chased(normal_ratios)

plot(stats::ecdf(chased_ratios))
is_sales_chased(chased_ratios)
}
