% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_type}
\alias{is_type}
\title{Safely determine variable type}
\usage{
is_type(subvar, class)
}
\arguments{
\item{subvar}{Name of submission variable}

\item{class}{Class of submission variable}
}
\value{
logical: TRUE if 'subvar' is of type 'class'
}
\description{
Safely determine variable type
}
