% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{ki_mki}
\alias{ki_mki}
\alias{ki_mki_met}
\title{Calculate Kakwani and Modified Kakwani Index}
\usage{
ki_mki(assessed, sale_price, na.rm = FALSE, mki = TRUE)

ki_mki_met(x)
}
\arguments{
\item{assessed}{vector or row of assessed values
with same length as \code{sale_price}}

\item{sale_price}{vector or row of sale values
with same length as \code{assessed}}

\item{na.rm}{Default FALSE. A boolean value indicating whether or not to
remove NA values. If missing values are present but not removed the
function will output NA.}

\item{mki}{Default TRUE. A boolean indicator to identify if the output
is mki or ki. MKI is the preferred metric.}

\item{x}{Numeric vector of sales ratio statistic(s) to check
against IAAO standards.}
}
\description{
The Kakwani Index (ki) and the Modified Kakwani Index (mki) are GINI-based measures
to test for vertical equity. The method first orders the housing stock by
increasing sale prices, and then calculates the GINI coefficient for sale value and
assessed value (remaining ordered by sales price). The Kakwani Index then
calculates the difference (GINI of Assessed - GINI of Sale), and the
Modified Kakwani Index calculates the ratio (GINI of Assessed / GINI of Sale).

KI < 0 is regressive
KI = 0 is vertical equity
KI > 0 is progressive

MKI < 1 is regressive
MKI = 1 is vertical equity
MKI > 1 is progressive

  \href{https://researchexchange.iaao.org/jptaa/vol17/iss2/2/}{A Gini measure for vertical equity in property assessments}
}
\section{Functions}{
\itemize{
\item \code{ki_mki()}: returns either the MKI or KI of numeric indicators
based on boolean indicator.

\item \code{ki_mki_met()}: Returns TRUE when input meets report standards.

}}
\examples{
# example code
result <- ki_mki(ratios_sample$assessed, ratios_sample$sale_price, mki = TRUE)

}
\seealso{
Other formulas: 
\code{\link{cod}()},
\code{\link{prb}()},
\code{\link{prd}()}
}
\concept{formulas}
