% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lgl_vecs_equal}
\alias{lgl_vecs_equal}
\title{Are logical vectors equal?}
\usage{
lgl_vecs_equal(subvar, sol_env, solvar = subvar, add = TRUE)
}
\arguments{
\item{subvar}{name of variable in submission environment}

\item{sol_env}{solution environment}

\item{solvar}{name of solution variable}

\item{add}{whether to add feedback}
}
\value{
logical
}
\description{
Are logical vectors equal?
}
