% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_compiled}
\alias{list_compiled}
\title{List all compiled html files}
\usage{
list_compiled(subdir, with_moodle_id = FALSE)
}
\arguments{
\item{subdir}{subdirectory in which to scan for html files}

\item{with_moodle_id}{extract the moodle identifier}
}
\value{
if \code{with_moodle_id} is \code{FALSE}, returns a vector
  with paths to submission files; if \code{TRUE}, a table with
  submission id and filename.
}
\description{
List all html files compiled from RMarkdown
}
