% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{list_submissions}
\alias{list_submissions}
\title{List all RMarkdown submissions}
\usage{
list_submissions(subdir, with_moodle_id = FALSE)
}
\arguments{
\item{subdir}{subdirectory in which to scan for RMarkdown files}

\item{with_moodle_id}{extract the moodle identifier}
}
\value{
if \code{with_moodle_id} is \code{FALSE}, returns a vector
  with paths to submission files; if \code{TRUE}, a table with
  submission id and filename.
}
\description{
List all RMarkdown submissions in a given subdirectory
}
