% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lmerMods_identical}
\alias{lmerMods_identical}
\title{Are lmer models identical?}
\usage{
lmerMods_identical(subvar, solenv, solvar = subvar, add = TRUE)
}
\arguments{
\item{subvar}{name of submission variable}

\item{solenv}{solution environment}

\item{solvar}{name of solution variable}

\item{add}{whether to add feedback}
}
\value{
logical
}
\description{
Are lmer models identical?
}
