% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lms_identical}
\alias{lms_identical}
\title{Are linear models identical?}
\usage{
lms_identical(subvar, solenv, solvar = subvar, add = TRUE)
}
\arguments{
\item{subvar}{name of variable in submission environment}

\item{solenv}{solution environment}

\item{solvar}{name of variable in solution environment}

\item{add}{whether to add feedback}
}
\value{
logical value
}
\description{
Are linear models identical?
}
