% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(obs, pred)
}
\arguments{
\item{obs}{T X 1 vector of observations.}

\item{pred}{T X N matrix of predictions where each column is a simulation.}
}
\value{
T X 1 vector of mean absolute error
}
\description{
Mean absolute error
}
\author{
Sangeeta Bhatia
}
