% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_same_dims}
\alias{matrix_same_dims}
\title{Do matrices have same dimensions?}
\usage{
matrix_same_dims(submx, sol_env, solmx = submx, add = TRUE)
}
\arguments{
\item{submx}{Quoted name of submission matrix.}

\item{sol_env}{Solution environment.}

\item{solmx}{Name of solution matrix.}

\item{add}{Whether to add feedback.}
}
\value{
Logical corresponding to whether matrices have same dimensions.
}
\description{
Do matrices have same dimensions?
}
