% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_vals_close}
\alias{matrix_vals_close}
\title{Are the matrix values close?}
\usage{
matrix_vals_close(submx, sol_env, solmx = submx, tolerance = 0.002,
  add = TRUE)
}
\arguments{
\item{submx}{Quoted name of submission matrix.}

\item{sol_env}{Solution environment.}

\item{solmx}{Name of solution matrix.}

\item{tolerance}{How close the values need to be to be considered identical.}

\item{add}{Whether to add feedback.}
}
\value{
Number of cell values matching the solution.
}
\description{
Are the matrix values close?
}
\details{
Note that the dimensions are not checked, so can return TRUE if the values are close but come from a matrix of a different dimension.
}
\seealso{
\code{\link{matrix_same_dims}}
}
