% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{med_ratio_met}
\alias{med_ratio_met}
\title{Check if Sales Ratio Meets IAAO Standards}
\usage{
med_ratio_met(x)
}
\arguments{
\item{x}{Numeric vector of sales ratios.}
}
\value{
Logical vector indicating whether each value meets the standard.
}
\description{
This function checks whether the input sales ratio falls within the IAAO standard range (0.9 to 1.1).
}
