% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{mki}
\alias{mki}
\alias{ki}
\alias{mki_met}
\title{Calculate Kakwani and Modified Kakwani Index}
\usage{
mki(assessed, sale_price, na.rm = FALSE)

ki(assessed, sale_price, na.rm = FALSE)

mki_met(x)
}
\arguments{
\item{assessed}{A numeric vector of assessed values. Must be the same
length as \code{sale_price}.}

\item{sale_price}{A numeric vector of sale prices. Must be the same length
as \code{assessed}.}

\item{na.rm}{Default FALSE. A boolean value indicating whether or not to
remove NA values. If missing values are present but not removed the
function will output NA.}

\item{x}{Numeric vector of sales ratio statistic(s) to check
against IAAO/accepted standards.}
}
\description{
The Kakwani Index (KI) and the Modified Kakwani Index (MKI)
are Gini-based methods to measure vertical equity.

These methods first order properties by sale price (ascending), then
calculate the Gini coefficient for sale values and assessed values (while
remaining ordered by sale price). The Kakwani Index then
calculates the difference \code{(Gini of assessed - Gini of sale)}, and the
Modified Kakwani Index calculates the ratio
\code{(Gini of Assessed / Gini of Sale)}.

For the Kakwani Index:
\itemize{
\item KI < 0 is regressive
\item KI = 0 is vertical equity
\item KI > 0 is progressive
}

For the Modified Kakwani Index:
\itemize{
\item MKI < 1 is regressive
\item MKI = 1 is vertical equity
\item MKI > 1 is progressive
}
}
\section{Functions}{
\itemize{
\item \code{mki()}: Returns a numeric vector containing the MKI of the
input vectors.

\item \code{ki()}: Returns a numeric vector containing the KI of the
input vectors.

\item \code{mki_met()}: Returns TRUE when input meets Quintos paper standards
(between 0.95 and 1.05).

}}
\examples{
# Calculate MKI
mki(ratios_sample$assessed, ratios_sample$sale_price)

# Calculate KI
ki(ratios_sample$assessed, ratios_sample$sale_price)
}
\references{
Quintos, C. (2020). A Gini measure for vertical equity in property
assessments. Journal of Property Tax Assessment & Administration, 17(2).
Retrieved from \href{https://researchexchange.iaao.org/jptaa/vol17/iss2/2}{https://researchexchange.iaao.org/jptaa/vol17/iss2/2}.

Quintos, C. (2021). A Gini decomposition of the sources of inequality in
property assessments. Journal of Property Tax Assessment & Administration,
18(2). Retrieved from
\href{https://researchexchange.iaao.org/jptaa/vol18/iss2/6}{https://researchexchange.iaao.org/jptaa/vol18/iss2/6}
}
\seealso{
Other formulas: 
\code{\link{cod}()},
\code{\link{prb}()},
\code{\link{prd}()}
}
\concept{formulas}
