% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moodle.R
\name{moodle_id}
\alias{moodle_id}
\title{Get Moodle identifier from filename}
\usage{
moodle_id(filename)
}
\arguments{
\item{filename}{Name of the file downloaded from Moodle}
}
\value{
Moodle submission identifier (integer)
}
\description{
Get Moodle identifier from filename
}

