% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlist_to_rmd}
\alias{nlist_to_rmd}
\title{Export named list as RMarkdown file}
\usage{
nlist_to_rmd(nlist, outfile = tempfile(), overwrite = FALSE)
}
\arguments{
\item{nlist}{A named list, with the names corresponding to chunk names for the output file.}

\item{outfile}{Name of the output file.}

\item{overwrite}{Whether to overwrite existing file.}
}
\value{
Path to output file, returned invisibly.
}
\description{
Export named list as RMarkdown file
}
