% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{num_vals_close}
\alias{num_vals_close}
\title{Are submission and solution numeric values close}
\usage{
num_vals_close(subvar, sol_env, solvar = subvar, ignore.case = FALSE,
  tolerance = 0.002, add = TRUE)
}
\arguments{
\item{subvar}{Name of the submission variable whose value you want to test}

\item{sol_env}{solution environment}

\item{solvar}{Name of the solution variable to test against}

\item{ignore.case}{whether to accept same variable name but different capitalization}

\item{tolerance}{how close the values have to be}

\item{add}{whether to add feedback}
}
\value{
logical; \code{TRUE} if \code{abs(x - get(x, sol_env)) < tolerance}
}
\description{
Test whether numeric values in the submission are close to corresponding values in the solution environment.
}

