% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{objs_identical}
\alias{objs_identical}
\title{Are objects identical}
\usage{
objs_identical(subobj, sol_env, solobj = subobj, add = TRUE)
}
\arguments{
\item{subobj}{name of submission object}

\item{sol_env}{solution environment}

\item{solobj}{name of solution object}

\item{add}{add feedback}
}
\value{
logical
}
\description{
Are objects identical
}
\details{
use this to compare any two objects (e.g., fitted model objects resulting from a call to `lm()`, `aov()`, etc)
}
