% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_funs.R
\name{pin_clean}
\alias{pin_clean}
\title{Clean input parameter values}
\usage{
pin_clean(x)
}
\arguments{
\item{x}{A Property Index Number (PIN) containing unnecessary whitespace
or separator characters}
}
\value{
A clean PIN character vector with no whitespace or separators.
}
\description{
This function removes common separators, whitespace, and trailing characters
from Property Index Numbers (PINs). It is specifically formatted for PINs,
which are expected to be 10 or 14 digit numbers saved as characters (to 
allow for leading zeros).
}
\examples{

pins <- c("04-34-106-008-0000", " 14172 27008 0000")

pin_clean(pins)

print(pins)

}
