% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{prb}
\alias{prb}
\title{Calculate Coefficient of Price-Related Bias (PRB)}
\usage{
prb(assessed, sale_price, na.rm = FALSE)
}
\arguments{
\item{assessed}{A numeric vector of assessed values. Must be the same
length as \code{sale_price}.}

\item{sale_price}{A numeric vector of sale prices. Must be the same length
as \code{assessed}.}

\item{na.rm}{Default FALSE. A boolean value indicating whether or not to
remove NA values. If missing values are present but not removed the
function will output NA.}
}
\value{
A numeric vector containing the PRB of the input vectors.
}
\description{
PRB is an index of vertical equity that quantifies the
  relationship betweem ratios and assessed values as a percentage. In
  concrete terms, a PRB of 0.02 indicates that, on average, ratios increase
  by 2% whenever assessed values increase by 100 percent.

  PRB is centered around 0 and has a generally accepted value of between
  -0.05 and 0.05, as defined in the
  \href{https://www.iaao.org/media/standards/Standard_on_Ratio_Studies.pdf}{IAAO Standard on Ratio Studies}
  Section 9.2.7. Higher PRB values indicate progressivity in assessment,
  while negative values indicate regressivity.

  NOTE: PRB is significantly less sensitive to outliers than PRD or COD.
}
\examples{

# Load the included dataset
data("ratios_sample")

# Calculate PRD
prb(ratios_sample$assessed, ratios_sample$sale_price)
}
\seealso{
Other formulas: 
\code{\link{cod}()},
\code{\link{prd}()}
}
\concept{formulas}
