% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assmnt_funs.R
\name{prb_func}
\alias{prb_func}
\title{Calculate Price-related Bias (PRB)}
\usage{
prb_func(
  ratios,
  sales,
  assessed_values,
  trim = c(0.05, 0.95),
  suppress = FALSE,
  na_rm = FALSE
)
}
\arguments{
\item{ratios}{A vector of ratios centered around 1, where the numerator of
the ratio is the estimated fair cash value and the denominator
is the sale price. NOTE: These values are typically lagged or leading
to prevent sales chasing.}

\item{sales}{A vector of sales the same length as \code{ratios}.}

\item{assessed_values}{A vector AVs the same length as \code{ratios}.}

\item{trim}{A lower and upper quantile for trimming input vectors. These
values are set by CCAO Data Science Department SOPs. Do not alter
them without permission.}

\item{suppress}{Default FALSE. If TRUE, hide warning about not meeting the
minimum number of observations to calculate a statistic accurately. The
minimum value of observation for accurate statistics is 30, per CCAO SOPs.}

\item{na_rm}{Default FALSE. If TRUE, will remove NAs from the vector before
proceeding.}
}
\value{
A named list containing the statistic, standard error,
  95\% confidence interval, and the number of observations used
  in the calculation.
}
\description{
Measure the PRB of a vector of assessment ratios and corresponding vectors
of assessment ratios and sales. Used by the CCAO to measure the
vertical equity of assessment models.
}
\details{
PRB is calculated using a regression method and has a closed-form solution
for calculating standard errors. As a result, there is no need for
bootstrapping.
}
\examples{
library(assessr)

# Load the included dataset
data("ratios_sample")

prb_func(
  ratios_sample$ratios,
  ratios_sample$sales,
  ratios_sample$assessed_values,
  trim = c(0.05, 0.95)
)

}
\seealso{
Other assmnt_functions: 
\code{\link{cod_func}()},
\code{\link{prd_func}()}
}
\concept{assmnt_functions}
