% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{prd}
\alias{prd}
\title{Calculate Price-Related Differential (PRD)}
\usage{
prd(assessed, sale_price, na.rm = FALSE)
}
\arguments{
\item{assessed}{A numeric vector of assessed values. Must be the same
length as \code{sale_price}.}

\item{sale_price}{A numeric vector of sale prices. Must be the same length
as \code{assessed}.}

\item{na.rm}{Default FALSE. A boolean value indicating whether or not to
remove NA values. If missing values are present but not removed the
function will output NA.}
}
\value{
A numeric vector containing the PRD of the input vectors.
}
\description{
PRD is the mean ratio divided by the mean ratio weighted by sale
  price. It is a measure of vertical equity in assessment. Vertical equity
  means that properties at different levels of the income distribution
  should be similarly assessed.

  PRD centers slightly above 1 and has a generally accepted value of between
  0.98 and 1.03, as defined in the
  \href{https://www.iaao.org/media/standards/Standard_on_Ratio_Studies.pdf}{IAAO Standard on Ratio Studies}
  Section 9.2.7. Higher PRD values indicate regressivity in assessment.

  NOTE: The IAAO recommends trimming outlier ratios before calculating PRD,
  as it is extremely sensitive to large outliers. PRD is being deprecated in
  favor of PRB, which is less sensitive to outliers and easier to interpret.
}
\examples{

# Load the included dataset
data("ratios_sample")

# Calculate PRD
prd(ratios_sample$assessed, ratios_sample$sale_price)
}
\seealso{
Other formulas: 
\code{\link{cod}()},
\code{\link{prb}()}
}
\concept{formulas}
