% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{prop_in_ci}
\alias{prop_in_ci}
\title{Proportion of observations in given credible interval}
\usage{
prop_in_ci(obs, min, max)
}
\arguments{
\item{obs}{vector of observed values}

\item{min}{vector of the lower end of the interval. Either
length 1 vector or the same length as the that of obs.}

\item{max}{vector of the upper end of the interval. Either length
1 or the same length as that of the obs vector.}
}
\value{
proportion of values in obs vector that are greater than
or equal to min and less than or equal to max.
}
\description{
Proportion of observations in given credible interval
}
\details{
Proportion of observed values that fall within a
given interval
}
\author{
Sangeeta Bhatia
}
