% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quintos_sample}
\alias{quintos_sample}
\title{Sample of sales and estimated market values provided by Quintos in the
   following MKI papers:}
\format{
A data frame with 30 observation and 2 variables:
\describe{
  \item{estimate}{Assessed fair market value}
  \item{sale_price}{Recorded sale price of this property}
}
}
\usage{
quintos_sample
}
\description{
Sample of sales and estimated market values provided by Quintos in the
   following MKI papers:
}
\references{
Quintos, C. (2020). A Gini measure for vertical equity in property
assessments. <https://researchexchange.iaao.org/jptaa/vol17/iss2/2>

Quintos, C. (2021). A Gini decomposition of the sources of inequality in
property assessments. <https://researchexchange.iaao.org/jptaa/vol18/iss2/6>
}
\keyword{datasets}
