% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratios_sample}
\alias{ratios_sample}
\title{Sample of ratio and sales data pulled from CCAO records.}
\format{
A data frame with 979 observation and 4 variables:
\describe{
  \item{fmv}{The fair market assessed value predicted by CCAO assessment
    models, including any successful appeals}
  \item{sale_price}{The recorded sale price of this property}
  \item{ratio}{Sales ratio representing fair market value / sale price}
  \item{town}{Township name the property is in}
}
}
\usage{
ratios_sample
}
\description{
This sample was take from Evanston and New Trier in 2019. Ratios are
calculated using assessor certified (post-appeal) fair market values.
}
\keyword{datasets}
