% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rel_mae}
\alias{rel_mae}
\title{Relative mean absolute error}
\usage{
rel_mae(obs, pred)
}
\arguments{
\item{obs}{T X 1 vector of observations.}

\item{pred}{T X N matrix of predictions where each column is a simulation.}
}
\value{
T X 1 vector of mean absolute error normalised by the observed value.
}
\description{
Relative mean absolute error
}
\details{
Relative mean absolute error is defined as
\deqn{\sum_{i = 1}^{N}{|obs - pred|} / N * |obs + 1|}
}
\author{
Sangeeta Bhatia
}
