% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rel_mean_dvtn}
\alias{rel_mean_dvtn}
\title{Relative sharpness}
\usage{
rel_mean_dvtn(pred)
}
\arguments{
\item{pred}{T X N Matrix of predictions. Each column is
a simulation.}
}
\value{
vector of length T.
}
\description{
Relative mean absolute deviation about the median
}
\details{
\deqn{median(|(pred - median(pred))/pred|)}
}
\references{
https://bit.ly/2vPO0I9
}
