% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rel_mse}
\alias{rel_mse}
\title{Average relative mean squared error}
\usage{
rel_mse(obs, pred)
}
\arguments{
\item{obs}{observed vector T X 1}

\item{pred}{matrix of predicted observations. Each column is a
simulation. T X N where N is the number of simulations.}
}
\value{
error T X 1. Each entry is the error averaged across the
simulations
}
\description{
Relative mean squared error averaged acorss simulations
}
\details{
Relative average mean square error is
\deqn{\sum_{i = 1}^{N}{(obs - pred)^2} / N * (obs + 1)^2}
We add 1 to the observed vector to avoid dividing by 0.
}
\author{
Sangeeta Bhatia
}
