% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{render_solutions}
\alias{render_solutions}
\title{Render dynamic solutions to R Markdown}
\usage{
render_solutions(outfile, solfile, task_vars = NULL, preamble = "",
  open = "{{", close = "}}", overwrite = FALSE)
}
\arguments{
\item{outfile}{File path of R Markdown output file.}

\item{solfile}{File path of generic solution file.}

\item{task_vars}{Named list of tasks, with sublists containing task
variables for the corresponding task. Task names should match
chunk names in the generic solution file.}

\item{preamble}{Character vector with lines to be written to the
solution before the first task chunk.}

\item{open}{Open delimiter for task variables.}

\item{close}{Close delimiter for task variables.}

\item{overwrite}{Overwrite existing output file.}
}
\value{
A character vector containing lines from the solution file, returned invisibly.
}
\description{
Create submission-specific solution files based on a generic
solution file and task variables.
}
\details{
Task variables are referenced in the solution file
  embedded between the open and close delimiters.
}
