% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_get_num}
\alias{safe_get_num}
\title{Safely get a double or an integer}
\usage{
safe_get_num(x, env = parent.frame(), inherits = FALSE, add = TRUE)
}
\arguments{
\item{x}{Quoted name of the variable.}

\item{env}{The environment in which to search.}

\item{inherits}{Whether to search in the parent environments.}

\item{add}{Whether or not to add feedback.}
}
\value{
A value of the desired type if found, or \code{NULL} if not found.
}
\description{
Safely get a double or an integer
}
\details{
First checks whether the variable exists in the environment. If it does, then checks whether it is of the appropriate type.
}
