% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{same_col_names}
\alias{same_col_names}
\title{Are table column names the same}
\usage{
same_col_names(subtbl, sol_env, soltbl = subtbl, add = TRUE)
}
\arguments{
\item{subtbl}{name of table in submission environment}

\item{sol_env}{Solution environment}

\item{soltbl}{name of table in solution environment}

\item{add}{whether to add feedback}
}
\value{
logical
}
\description{
Are table column names the same
}

