% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{score_mcq}
\alias{score_mcq}
\title{Score multiple-choice question}
\usage{
score_mcq(
  subvar,
  sol_env,
  opts_all,
  solvar = subvar,
  rectify = TRUE,
  add = TRUE
)
}
\arguments{
\item{subvar}{Name of the submission variable}

\item{sol_env}{Solution environment}

\item{opts_all}{Vector whose elements comprise the full set of options (e.g., "a", "b", "c", "d")}

\item{solvar}{Name of the solution variable}

\item{rectify}{Whether to attempt to correct the format of the submitted answer}

\item{add}{whether to add feedback}
}
\value{
vector of logicals, one for each option, plus "type" and "fmt_ok" to indicate correctly formatted answer
}
\description{
Score a multiple-choice question.
}
