% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{score_mcq_opt}
\alias{score_mcq_opt}
\title{Score a single multiple-choice option}
\usage{
score_mcq_opt(subvar, sol_env, opt, solvar = subvar, add = TRUE)
}
\arguments{
\item{subvar}{Name of the submission variable}

\item{sol_env}{Solution environment}

\item{opt}{Name of the option (e.g., "a", "b", "c")}

\item{solvar}{Name of the solution variable}

\item{add}{whether to add feedback}
}
\value{
logical; \code{TRUE} if element is present in both vectors or absent in both vectors
}
\description{
Checks  a single choice in the submission variable against the solution.
}
