% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{scrape_inline}
\alias{scrape_inline}
\title{Tangle inline code from file}
\usage{
scrape_inline(filename)
}
\arguments{
\item{filename}{name of RMarkdown file}
}
\value{
list with inline code only
}
\description{
Tangle inline code from file
}

