% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setextra}
\alias{setextra}
\title{Test set difference}
\usage{
setextra(x, y, v)
}
\arguments{
\item{x}{First set}

\item{y}{Second set}

\item{v}{Difference vector}
}
\value{
logical
}
\description{
Does the difference in elements between first set x and second set
y correspond to vector v?
}
