% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setextra_mod}
\alias{setextra_mod}
\title{Test difference in model (fixed effects) terms}
\usage{
setextra_mod(x, y, v)
}
\arguments{
\item{First}{model}

\item{Second}{model}

\item{Difference}{vector}
}
\value{
logical
}
\description{
Test difference in model (fixed effects) terms
}
\details{
See whether two models differ in fixed effects terms. This can be tricky because interaction terms that are ordered differently (\code{x:y} and \code{y:x}) should be treated as identical. Any interactions will always be returned sorted in alphabetical order.
}
