% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{tangle}
\alias{tangle}
\title{Tangle code from Rmd file}
\usage{
tangle(filename)
}
\arguments{
\item{filename}{name of Rmd file to tangle}
}
\value{
A list containing file blocks
}
\description{
Extract code blocks from RMarkdown file
}

