% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_cols_sameclass}
\alias{tbl_cols_sameclass}
\title{Are table column values equal}
\usage{
tbl_cols_sameclass(
  subtbl,
  subcol,
  sol_env,
  soltbl = subtbl,
  solcol = subcol,
  add = TRUE
)
}
\arguments{
\item{subtbl}{name of submission table}

\item{subcol}{name of column in submission table}

\item{sol_env}{solution environment}

\item{soltbl}{name of table of solution environment}

\item{solcol}{name of column in solution table}

\item{add}{whether to add feedback}
}
\value{
logical
}
\description{
Are table column values equal
}
