% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_has_same_nrows}
\alias{tbl_has_same_nrows}
\title{Same number of rows in table}
\usage{
tbl_has_same_nrows(subtbl, sol_env, soltbl = subtbl, add = TRUE)
}
\arguments{
\item{subtbl}{name of table in submission environment}

\item{sol_env}{solution environment}

\item{soltbl}{name of table in solution environment}

\item{add}{whether to add feedback}
}
\value{
logical
}
\description{
Same number of rows in table
}

