% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_same_dims}
\alias{tbl_same_dims}
\title{Are table dimensions the same}
\usage{
tbl_same_dims(subtbl, sol_env, soltbl = subtbl, add = TRUE)
}
\arguments{
\item{subtbl}{name of table in submission environment}

\item{sol_env}{Solution environment}

\item{soltbl}{name of table in solution environment}

\item{add}{whether to add feedback}
}
\value{
logical
}
\description{
Are table dimensions the same
}
