% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbls_identical}
\alias{tbls_identical}
\title{Are submission and solution tables identical}
\usage{
tbls_identical(subtbl, sol_env, soltbl = subtbl, ignore.case = FALSE,
  roworder_strict = FALSE, colorder_strict = FALSE,
  allow_extracols = FALSE, add = TRUE)
}
\arguments{
\item{subtbl}{Name of the table in submission environment.}

\item{sol_env}{The solution environment.}

\item{soltbl}{Name of the table in the solution environment.}

\item{ignore.case}{Whether to ignore case when matching submission and solution variable names.}

\item{roworder_strict}{Whether to require the submission table to have rows in the same order as the solution table.}

\item{colorder_strict}{Whether to require the submission table to have columns in the same order as the target table.}

\item{allow_extracols}{Whether to allow the submission table to have extra columns not present in the solution table.}

\item{add}{whether to add feedback}
}
\value{
logical; returns result of \code{dplyr::setequal(tblname, get(tblname, sol_env))}
}
\description{
Are submission and solution tables identical
}
