% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/town_funs.R
\name{town_convert}
\alias{town_convert}
\title{Convert between township number and township name}
\usage{
town_convert(town)
}
\arguments{
\item{town}{Vector of township names or numeric codes. Township name
must be title case to match correctly.}
}
\value{
Vector or township codes or names equivalent to the input vector.
  Returns NA for towns that cannot be found.
}
\description{
Convert township name to its respective two-digit code and visa-versa.
}
\examples{

town_convert("Evanston")

town_convert(c("31", "25"))

town_convert(c("31", "Proviso", "Lyons"))
}
