% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{town_dict}
\alias{town_dict}
\title{Data dictionary for CCAO township codes and triads.}
\format{
A data frame with 38 rows and 4 variables:
\describe{
  \item{township_name}{Common name of the township}
  \item{township_code}{Two-digit code used to identify the township}
  \item{triad_code}{Single-digit code of the triad the township is in}
  \item{triad_name}{Common name of the triad the township is in}
}
}
\usage{
town_dict
}
\description{
A dataset containing a lookup of various townships and their
respective triads.
}
\keyword{datasets}
