% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/town_funs.R
\name{town_get_assmnt_year}
\alias{town_get_assmnt_year}
\title{Get the nearest assessment year of a township}
\usage{
town_get_assmnt_year(town, year = as.integer(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{town}{Vector of township names or numeric codes. Township name
must be title case to match correctly.}

\item{year}{Default current year. Year to find assessment year closest to.
For example, entering 1995 will find the closest assessment year to 1995
for all the entered triads.}
}
\value{
Numeric vector of assessment years closest to the year entered.
  Returns NA for towns that cannot be found.
}
\description{
Input a township name or code and get the assessment year closest to the
year argument. Current year is the default. Will never return a year higher
than the current year.
}
\details{
Ex. \code{town_get_assmnt_year("Evanston")} outputs \code{2019}
}
\examples{

town_get_assmnt_year("Evanston")

town_get_assmnt_year(c("Lyons", "10", "Worth"), year = 1997)
}
