% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/town_funs.R
\name{town_get_triad}
\alias{town_get_triad}
\title{Get Cook County triad name or number based on township name or code}
\usage{
town_get_triad(town, name = FALSE)
}
\arguments{
\item{town}{Vector of township names or numeric codes. Township name
must be title case to match correctly.}

\item{name}{Default FALSE. Output triad name instead of triad number.}
}
\value{
Vector of triad numbers for the input townships. Returns NA for
  towns that cannot be found.
}
\description{
Get the triad name or number based on township name or code.
Cook County is split into three sections called triads. These triads are
assessed every three years. Each town lies entirely within a single triad.
}
\examples{

town_get_triad("Evanston")

town_get_triad(c("Lyons", "10", "Worth"))
}
