% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ttest_identical}
\alias{ttest_identical}
\title{Check whether t-test objects are identical}
\usage{
ttest_identical(subvar, sol_env, solvar = subvar, tolerance = c(0.2,
  0.02, 0.2, 0.002), add = TRUE)
}
\arguments{
\item{subvar}{Quoted name of the variable.}

\item{sol_env}{Solution environment.}

\item{solvar}{Quoted name of variable in submission environment.}

\item{tolerance}{Three-element numeric vector, how close values have to be.}

\item{add}{Whether to add feedback.}
}
\value{
A logical vector with elements mmatch, tmatch, dfmatch, pmatch.
}
\description{
Check whether t-test objects are identical
}
