% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_vals_close}
\alias{vec_vals_close}
\title{Are vector element values close?}
\usage{
vec_vals_close(subvec, sol_env, solvec = subvec, tolerance = 0.002,
  add = TRUE)
}
\arguments{
\item{subvec}{name of the vector in the submission environment}

\item{sol_env}{the solution environment}

\item{tolerance}{how close the values have to be}

\item{add}{whether to add feedback}
}
\value{
logical
}
\description{
Are vector element values close?
}
