% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{visit}
\alias{visit}
\title{Visit a submission file}
\usage{
visit(files, x)
}
\arguments{
\item{files}{a vector with names of submission files}

\item{x}{moodle id number (see \code{\link{moodle_id}})}
}
\description{
Visit a submission file
}
\details{
Finds the filename within \code{files} that corresponds to
  a given moodle id, and opens it with \code{file.edit}
}
