% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assess.R
\name{add_feedback}
\alias{add_feedback}
\title{Add feedback to end of code block}
\usage{
add_feedback(..., sep = "", add = TRUE)
}
\arguments{
\item{...}{character strings giving feedback}

\item{sep}{separator for character strings}

\item{add}{whether to add or ignore}
}
\description{
Add feedback to end of code block
}

