% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{apply_blacklist}
\alias{apply_blacklist}
\title{Check for blacklisted functions}
\usage{
apply_blacklist(subfile, overwrite = TRUE)
}
\arguments{
\item{subfile}{path to submission RMarkdown file}

\item{overwrite}{if you want to comment out the lines from the original file}
}
\value{
a tibble with columns \code{blacklist}, whether any functions were found; and \code{bfns}, names of these functions
}
\description{
Search for, and optionally comment out any blacklisted functions
that appear within an RMarkdown file
}
\details{
warning: the original file will be overwritten, with any
  lines containing blacklisted code commented out
}
\seealso{
\code{\link{get_blacklist}}
}
