% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess_all}
\alias{assess_all}
\title{Assess all files in a directory}
\usage{
assess_all(dirname, key, sub_id = moodle_id(list_submissions(dirname)),
  use_sub_env = TRUE, workdir = NULL, seed = NULL, stop_after = -1L)
}
\arguments{
\item{dirname}{name of the directory containing the submission files}

\item{key}{key (answers to the problems); normally result of call to \code{link{compile_key}}}

\item{sub_id}{subject identifies (vector same length as \code{dirname})}

\item{use_sub_env}{process submission code in the submission environment (\code{FALSE} to process it in the solution environment)}

\item{workdir}{working directory}

\item{seed}{random seed}

\item{stop_after}{stop processing after completing N files}
}
\value{
a dataframe with the assessment variables and their values
}
\description{
Convenience function that runs the function \code{assess} on all files within the directory specified by \code{dirname}.
}
