% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assess.R
\name{assess_task}
\alias{assess_task}
\title{Assess individual task}
\usage{
assess_task(sub_id, task, sub_code, a_code, orig_env)
}
\arguments{
\item{sub_id}{}

\item{task}{}

\item{sub_code}{}

\item{a_code}{}

\item{orig_env}{}
}
\description{
Assess an individual task from a submission
}

