% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess_task}
\alias{assess_task}
\title{Assess individual task}
\usage{
assess_task(sub_id, task, codelist, a_code, orig_env, sol_env,
  use_sub_env = TRUE)
}
\arguments{
\item{sub_id}{submission id}

\item{task}{name of task to assess}

\item{codelist}{list of all submission code chunks}

\item{a_code}{assessment code}

\item{orig_env}{starting environment in which to evaluate submission code}

\item{use_sub_env}{set to \code{TRUE} if you want to run in a single submission environment; \code{FALSE} to use the solution as starting environment}
}
\description{
Assess an individual task from a submission
}
