% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{chr_vecs_equal}
\alias{chr_vecs_equal}
\title{Are character vectors equal?}
\usage{
chr_vecs_equal(subvar, sol_env, solvar = subvar, ignore_order = FALSE,
  add = TRUE)
}
\arguments{
\item{subvar}{name of variable in submission environment}

\item{sol_env}{solution environment}

\item{solvar}{name of solution variable}

\item{ignore_order}{whether to ignore the ordering}

\item{add}{whether to add feedback}
}
\value{
logical
}
\description{
Are character vectors equal?
}

