% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{cod}
\alias{cod}
\title{Calculate Coefficient of Dispersion (COD)}
\usage{
cod(ratio, na.rm = FALSE)
}
\arguments{
\item{ratio}{A numeric vector of ratios centered around 1, where the
numerator of the ratio is the estimated fair market value and the
denominator is the actual sale price.}

\item{na.rm}{Default FALSE. A boolean value indicating whether or not to
remove NA values. If missing values are present but not removed the
function will output NA.}
}
\value{
A numeric vector containing the COD of \code{ratios}.
}
\description{
COD is the average absolute percent deviation from the
  median ratio. It is a measure of horizontal equity in assessment.
  Horizontal equity means properties with a similar fair market value
  should be similarly assessed.

  Lower COD indicates higher uniformity/horizontal equity in assessment.
  The IAAO sets uniformity standards that define generally accepted ranges
  for COD depending on property class. See
  \href{https://www.iaao.org/media/standards/Standard_on_Ratio_Studies.pdf}{IAAO Standard on Ratio Studies}
  Section 9.1, Table 1.3 for a full list of standard COD ranges.

  NOTE: The IAAO recommends trimming outlier ratios before calculating COD,
  as it is extremely sensitive to large outliers. The typical method used is
  dropping values beyond 3 * IQR (inner-quartile range). See
  \href{https://www.iaao.org/media/standards/Standard_on_Ratio_Studies.pdf}{IAAO Standard on Ratio Studies}
  Appendix B.1.
}
\examples{

# Load the included dataset
data("ratios_sample")

# Calculate COD
cod(ratios_sample$ratio)
}
\seealso{
Other formulas: 
\code{\link{prb}()},
\code{\link{prd}()}
}
\concept{formulas}
