% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assmnt_funs.R
\name{cod_func}
\alias{cod_func}
\title{Calculate Coefficient of Dispersion (COD)}
\usage{
cod_func(
  ratios,
  trim = c(0.05, 0.95),
  bootstrap_n = 100,
  suppress = FALSE,
  na_rm = FALSE
)
}
\arguments{
\item{ratios}{A vector of ratios centered around 1, where the numerator of
the ratio is the estimated fair cash value and the denominator
is the sale price. NOTE: These values are typically lagged or leading
to prevent sales chasing.}

\item{trim}{A lower and upper quantile for trimming input vectors. These
values are set by CCAO Data Science Department SOPs. Do not alter
them without permission.}

\item{bootstrap_n}{The number of iterations to use to estimate the output
statistic, standard error, and 95\% confidence interval.
Setting to 0 or FALSE will not use bootstrapping.}

\item{suppress}{Default FALSE. If TRUE, hide warning about not meeting the
minimum number of observations to calculate a statistic accurately. The
minimum value of observation for accurate statistics is 30, per CCAO SOPs.}

\item{na_rm}{Default FALSE. If TRUE, will remove NAs from the vector before
proceeding.}
}
\value{
A named list containing the statistic, standard error,
  95\% confidence interval, and the number of observations used
  in the calculation.
}
\description{
Measure the COD of a vector of assessment ratios. Used by the CCAO to
measure the uniformity of assessment models.
}
\examples{
library(assessr)

# Load the included dataset
data("ratios_sample")

cod_func(ratios_sample$ratios, trim = c(0.05, 0.95), bootstrap_n = 100)
}
\seealso{
Other assmnt_functions: 
\code{\link{prb_func}()},
\code{\link{prd_func}()}
}
\concept{assmnt_functions}
