% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{compile_all}
\alias{compile_all}
\title{Compile all RMarkdown submissions}
\usage{
compile_all(subdir, quiet = TRUE, blacklist = TRUE, progress = TRUE,
  with_moodle_id = FALSE)
}
\arguments{
\item{subdir}{subdirectory containing RMarkdown submissions}

\item{quiet}{suppress messages while compiling}

\item{blacklist}{comment out lines of code containing blacklisted functions}

\item{progress}{report file-by-file progress}

\item{return}{should return table contain moodle id number}
}
\value{
table with information about compilation for each submission
}
\description{
Compile all RMarkdown submissions
}
\details{
Convenience function to call \code{\link{compile_assignment}} repeatedly for each RMarkdown file in subdirectory \code{subdir}
}

