% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{compile_assignment}
\alias{compile_assignment}
\title{Compile assignment report}
\usage{
compile_assignment(subfile, quiet = FALSE, blacklist = TRUE)
}
\arguments{
\item{subfile}{path to submission RMarkdown file}

\item{quiet}{whether to run quietly (suppress messages)}

\item{blacklist}{comment out lines of code containing blacklisted functions}
}
\description{
Safely compile assignment report from submitted RMarkdown script
}
\details{
runs a safe version of \code{\link{rmarkdown::render}} on
  the submision file
}

