% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{compile_key}
\alias{compile_key}
\title{Compile the assessment key}
\usage{
compile_key(
  s_file,
  a_file,
  overwrite = FALSE,
  workdir = NULL,
  save_fig = TRUE,
  preseed = NULL
)
}
\arguments{
\item{s_file}{name of Rmd solution file}

\item{a_file}{name of Rmd file with assessment code}

\item{overwrite}{overwrite \code{solution} directory if it exists?}

\item{workdir}{working directory}

\item{preseed}{Named list of seeds to be passed to \code{set.seed} before running each solution code chunk. The element names should match the assessment and solution code chunk names.}

\item{savefig}{save any files created in solution subdirectory?}
}
\value{
a list containing the starting environments for each task
}
\description{
Compile the assessment key, creating starting and solution
environments to use in assessment, as well as output to use in
feedback reports.
}
\details{
Sets up environments for running submission code and
  assessment code.  Also creates output/plots when appropriate and
  stores them in the \code{solution} subdirectory in the current
  working directory.
}
\seealso{
\code{\link{tangle}}
}
