% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defined_with_code}
\alias{defined_with_code}
\title{Test whether a variable was defined statically or using code}
\usage{
defined_with_code(x, code, add = TRUE)
}
\arguments{
\item{x}{name of the variable to check}

\item{add}{whether to add feedback}
}
\value{
logical; \code{TRUE} if the code defining the variable uses any function
}
\description{
Test whether a variable was defined statically or using code
}
