% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{is_outlier}
\alias{is_outlier}
\alias{quantile_outlier}
\alias{iqr_outlier}
\title{Detect Outlier Values}
\usage{
is_outlier(x, method = "iqr", na.rm = FALSE, ...)

quantile_outlier(x, probs = c(0.05, 0.95), na.rm = FALSE)

iqr_outlier(x, mult = 3, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector. Must be longer than 2 and not contain
\code{Inf} or \code{NaN}.}

\item{method}{String indicating outlier detection method. Options are
\code{iqr} or \code{quantile}.}

\item{na.rm}{Default FALSE. A boolean value indicating whether or not to
remove NA values. If missing values are present but not removed the
function will output NA for those values.}

\item{...}{Named arguments passed on to methods.}

\item{probs}{Upper and lower percentiles denoting outlier boundaries.}

\item{mult}{Multiplier for IQR to determine outlier boundaries.}
}
\value{
A logical vector this same length as \code{x} indicating whether or
  not each value of \code{x} is an outlier.
}
\description{
Detect outliers in a numeric vector using standard methods.

  Certain assessment performance statistics are sensitive to extreme
  outliers. As such, it is often necessary to remove outliers before
  performing a sales ratio study.

  Standard method is to remove outliers that are 3 * IQR. Warnings are thrown
  when sample size is extremely small or when the IQR is extremely narrow. See
  \href{https://www.iaao.org/media/standards/Standard_on_Ratio_Studies.pdf}{IAAO Standard on Ratio Studies}
  Appendix B. Outlier Trimming Guidelines for more information.
}
\section{Functions}{
\itemize{
\item \code{quantile_outlier}: Quantile method for identifying outliers.

\item \code{iqr_outlier}: IQR method for identifying outliers.
}}

