% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{ki_mki}
\alias{ki_mki}
\alias{ki_mki_met}
\title{Calculate Kakwani and Modified Kakwani Index}
\usage{
ki_mki(assessed, sale_price, na.rm = FALSE, mki = TRUE)

ki_mki_met(x)
}
\arguments{
\item{assessed}{vector or row of assessed values
with same length as \code{sale_price}}

\item{sale_price}{vector or row of sale values
with same length as \code{assessed}}

\item{na.rm}{Default FALSE. A boolean value indicating whether or not to
remove NA values. If missing values are present but not removed the
function will output NA.}

\item{mki}{Boolean indicator to identify if the output is mki or ki.
mki = TRUE and default
ki = False}

\item{x}{Numeric vector of sales ratio statistic(s) to check
against IAAO standards.}
}
\description{
The GINI Measure for Vertical Equity is a mechanism to identify
difference in inequality for assessed and sale value of houses.
It uses two instances of the GINI coefficient, one for the distribution
of sold housing prices, and then assessed prices which remain sorted by sale price.
From these coefficients, two metrics are produced. 1st is the Kakwani Index (KI),
GINI of the assessed value - the GINI of the sale value.
2nd is the Modified Kakwani Index (MKI),
GINI of the assessed value / the GINI of the sale value.
The Modified Kakwani index is the standard, and a better metric when working
across multiple neighborhoods, with different distributions of sale values.

KI < 0 is regressive
KI = 0 is vertical equity
KI > 0 is progressive

MKI < 1 is regressive
MKI = 1 is vertical equity
MKI > 1 is progressive

  \href{https://researchexchange.iaao.org/jptaa/vol17/iss2/2/}{A Gini measure for vertical equity in property assessments}
}
\section{Functions}{
\itemize{
\item \code{ki_mki()}: returns a list of two results, the MKI and the KI.

\item \code{ki_mki_met()}: Returns TRUE when input meets report standards.

}}
\examples{
# example code
result <- ki_mki(ratios_sample$assessed, ratios_sample$sale_price, mki = TRUE)

}
\seealso{
Other formulas: 
\code{\link{cod}()},
\code{\link{prb}()},
\code{\link{prd}()}
}
\concept{formulas}
