% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{objs_identical}
\alias{objs_identical}
\title{Are objects the same}
\usage{
objs_identical(subobj, sol_env, solobj = subobj, add = TRUE)

objs_identical(subobj, sol_env, solobj = subobj, add = TRUE)
}
\arguments{
\item{subobj}{name of submission object}

\item{sol_env}{solution environment}

\item{solobj}{name of solution object}

\item{add}{add feedback}

\item{all_equal}{whether to use \code{all_equal} instead of \code{identical}}
}
\value{
logical
}
\description{
Are objects the same

Are objects equal
}
\details{
\code{objs_identical} uses \code{identical}; \code{objs_all_equal} usesuse this to compare any two objects (e.g., fitted model objects resulting from a call to `lm()`, `aov()`, etc)
}
\section{Functions}{
\itemize{
\item \code{objs_identical}: 
}}

