% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iaao_funs.R
\name{prb_func}
\alias{prb_func}
\title{Calculate Price-related Bias (PRB)}
\usage{
prb_func(
  ratios,
  assessed_values,
  sales,
  trim = c(0.05, 0.95),
  bootstrap_n = 100,
  suppress = FALSE,
  na_rm = FALSE
)
}
\arguments{
\item{ratios}{A vector of ratios centered around 1, where the numerator of
the ratio is the estimated fair cash value and the denominator
is the sale price. NOTE: These values are typically lagged or leading
to prevent sales chasing.}

\item{assessed_values}{A vector AVs the same length as \code{ratios}.}

\item{sales}{A vector of sales the same length as \code{ratios}.}

\item{trim}{A lower and upper quantile for trimming input vectors. These
values are set by CCAO Data Science Department SOPs. Do not alter
them without permission.}

\item{bootstrap_n}{The number of iterations to use to estimate the output
statistic, standard error, and 95\% confidence interval.
Setting to 0 or FALSE will not use bootstrapping.}

\item{suppress}{Default FALSE. If TRUE, hide warning about not meeting the
minimum number of observations to calculate a statistic accurately. The
minimum value of observation for accurate statistics is 30, per CCAO SOPs.}

\item{na_rm}{Default FALSE. If TRUE, will remove NAs from the vector before
proceeding.}
}
\value{
A named list containing the statistic, standard error,
  95\% confidence interval, and the number of observations used
  in the calculation.
}
\description{
Measure the PRB of a vector of assessment ratios and corresponding vectors
of assessment ratios and sales. Used by the CCAO to measure the
vertical equity of assessment models.
}
\examples{
library(assessR)

ratios <- runif(300, 0.93, 1.08)
assessed_values <- runif(300, 100000, 1000000)
sales <- runif(300, 100000, 1000000)

prb_func(ratios, sales, assessed_values)
\dontrun{
# With pre-existing vectors of ratios, AVs, and sales, stored in df

prb_func(df$ratio_6, df$fitted_value_6, df$sales, trim = c(0.05, 0.95))
}

}
\seealso{
Other iaao_functions: 
\code{\link{cod_func}()},
\code{\link{prd_func}()}
}
\concept{iaao_functions}
