% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R, R/ci.R
\name{prd}
\alias{prd}
\alias{prd_ci}
\alias{prd_met}
\title{Calculate Price-Related Differential (PRD)}
\usage{
prd(estimate, sale_price, na.rm = FALSE)

prd_ci(estimate, sale_price, nboot = 1000, alpha = 0.05, na.rm = FALSE)

prd_met(x)
}
\arguments{
\item{estimate}{A numeric vector of assessed values. Must be the same
length as \code{sale_price}.}

\item{sale_price}{A numeric vector of sale prices. Must be the same length
as \code{estimate}.}

\item{na.rm}{Default FALSE. A boolean value indicating whether or not to
remove NA values. If missing values are present but not removed the
function will output NA.}

\item{nboot}{Default 1000. Number of iterations to use to estimate
the output statistic confidence interval.}

\item{alpha}{Default 0.05. Numeric value indicating the confidence
interval to return. 0.05 will return the 95\% confidence interval.}

\item{x}{Numeric vector of sales ratio statistic(s) to check
against IAAO/Quintos standards.}
}
\description{
PRD is the mean ratio divided by the mean ratio weighted by sale
  price. It is a measure of vertical equity in assessment. Vertical equity
  means that properties at different levels of the income distribution
  should be similarly assessed.

  PRD centers slightly above 1 and has a generally accepted value of between
  0.98 and 1.03, as defined in the
  \href{https://www.iaao.org/media/standards/Standard_on_Ratio_Studies.pdf}{IAAO Standard on Ratio Studies}
  Section 9.2.7. Higher PRD values indicate regressivity in assessment.

  NOTE: The IAAO recommends trimming outlier ratios before calculating PRD,
  as it is extremely sensitive to large outliers. PRD is being deprecated in
  favor of PRB, which is less sensitive to outliers and easier to interpret.
}
\section{Functions}{
\itemize{
\item \code{prd()}: Returns a numeric vector containing the PRD of the
input vectors.

\item \code{prd_ci()}: Returns upper and lower CI as a named vector.

\item \code{prd_met()}: Returns TRUE when input PRD meets IAAO standards
(between 0.98 and 1.03).

}}
\examples{
# Calculate PRD
prd(ratios_sample$estimate, ratios_sample$sale_price)

# Calculate PRD confidence interval
prd_ci(ratios_sample$estimate, ratios_sample$sale_price)
}
\seealso{
Other formulas: 
\code{\link{cod}()},
\code{\link{mki}()},
\code{\link{prb}()}
}
\concept{formulas}
