% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cknn.R
\name{predict.cknn}
\alias{predict.cknn}
\title{Predict cluster and k nearest neighbors for new properties}
\usage{
\method{predict}{cknn}(object, newdata, lon, lat, k = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{"cknn"}.}

\item{newdata}{A data frame containing data in the same format as the
data used to create \code{obj}, but with new observations.}

\item{lon}{A numeric vector of longitude values, in the same projection as
the values stored in \code{obj}.}

\item{lat}{A numeric vector of latitude values, in the same projection as
the values stored in \code{obj}.}

\item{k}{An optional value for the number nearest neighbors to return
for each row of new data. Set to value used to create
\code{obj} by default.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
List containing:

\item{cluster}{Numeric vector of cluster membership for each row
  in the new data.}

\item{knn}{List of \code{k} nearest neighbors for each row in the
  new data.}
}
\description{
Predict cluster membership and \code{k} nearest neighbors
  for data not included in the original \code{\link{cknn}} function. See the
  \href{https://ccao-data-science---modeling.gitlab.io/packages/assessr/articles/example-ratio-study.html}{documentation site}
  for example usage.
}
\seealso{
Other cknn: 
\code{\link{cknn}()}
}
\concept{cknn}
