% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratios_sample}
\alias{ratios_sample}
\title{Sample of ratio and sales data used by the CCAO to evaluate assessment
performance.}
\format{
A data frame with 2979 observation and 4 variables:
\describe{
  \item{ratios}{Sales ratio representing assessed value / sale price}
  \item{sales}{The recorded sale price of this property}
  \item{town}{Township name the property is in}
  \item{assessed_values}{The value predicted by CCAO assessment models,
  including any appeals have been applied}
}
}
\usage{
ratios_sample
}
\description{
This sample was take from Evanston and New Trier in 2019. Ratios are
calculated using assessor certified (post-appeal) values.
}
\keyword{datasets}
