% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rmd.R
\name{tangle}
\alias{tangle}
\title{Tangle code from Rmd file}
\usage{
tangle(filename, inline_code = FALSE, documentation = 1L)
}
\arguments{
\item{filename}{name of Rmd file to tangle}

\item{inline_code}{keep inline code?}

\item{documentation}{see \code{\link{knitr::purl}}}
}
\value{
A list containing file blocks
}
\description{
Extract code blocks from RMarkdown file
}
