% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbls_identical}
\alias{tbls_identical}
\title{Are submission and solution tables identical}
\usage{
tbls_identical(subtbl, sol_env, soltbl = subtbl, ignore.case = FALSE,
  add = TRUE)
}
\arguments{
\item{subtbl}{name of the table in submission environment}

\item{sol_env}{the solution environment}

\item{soltbl}{name of the table in the solution environment}

\item{ignore.case}{whether to ignore case when matching submission and solution variable names}

\item{add}{whether to add feedback}
}
\value{
logical; returns result of \code{dplyr::setequal(tblname, get(tblname, sol_env))}
}
\description{
Are submission and solution tables identical
}
